backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
cloudscraper==1.2.71
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
diskcache==5.6.3
pymongo==4.6.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-multipart==0.0.6